[PowerPC] Support powerpc64le as a syntax-checking target.
This patch modifies the config.guess file to do the appropriate thing
on a system identifying itself as powerpc64le Linux.  As there
currently is no such system, there is nothing to test.  This is done
for completeness to accompany the LLVM and Clang patches to support
powerpc64le as a syntax-checking target.

llvm-svn: 187181