[PPC64LE] Fix Altivec 'test1.c' test to be endian-tolerant
The test1.c test in the SingleSource/UnitTests/Vector/Altivec
directory makes use of the vec_sld interface.  Unlike most of the
vector interfaces, vec_sld's behavior is not automatically modified
by the compiler for little endian.  (A programmer might use it to
multiply by a power of 2, for example, in which case it should not be
changed.)  Thus code using vec_sld must sometimes be modified when
porting to little endian.  This patch replaces two of the vec_sld
calls with replacement code appropriate for little endian.  A third
instance is correct for both big and little endian and doesn't need to
be modified.  A fourth instance is dead code and therefore doesn't
need to be modified.

llvm-svn: 210658