[test-suite] Add frame_layout test to better check correctness of frame lowering.
The added test in this patch uses a template function, so that depending on the
template parameters, the different areas on the stack (arguments, locals, VLAs,
alignment spacing, spill slots, ...) are either present or not present. By
instantiating this function for all combination of its template parameters, the
basic correct setup of the frame for all combinations of frame areas being
present or not, is being tested.

The test case needs some C++11 features - so this is the first test added to
the test suite that relies on C++11. I've created a separate sub-directory for
tests needing C++11 features, so hopefully it'll be possible for people to
still use the test-suite with compilers not supporting C++11, if that would be
needed.

llvm-svn: 236085