Cleanup and generalize .test file creation
This adds flexibility to the cmake/lit system to handle things like SPEC
in the future.  This change allows to directly specify the resulting
RUN: and VERIFY: lines in the .test file with llvm_test_run() and
llvm_test_verify() in the cmake file instead of relying on RUN_OPTIONS,
STDIN_FILENAME, HASH_PROGRAM_OUTPUTS, etc.

As discussed I am going for post-commit review as cmake/lit test-suite
support is still in early development.

Differential Revision: http://reviews.llvm.org/D14678

llvm-svn: 254832