[test-suite] Prevent timeit/fpcmp from being rebuilt every time make is run.
This patch adds a custom_command level underneath the custom_targets for timeit
and fpcmp to prevent the executables from being rebuilt every time make is run.

The custom_target rule will still be run every time, but the binaries will be
detected as being up to date and skip rebuilding.

http://reviews.llvm.org/D15938

llvm-svn: 257059