Add --param=profile=perf option to lit.
This option will also run each test under perf record, and store the
perf data in the Output directory.

At least on my system, there seems to be a limitation on running many
perf profilers in parallel, so for best results it's probably best to
use lit -j1 in combination with this option.

Differential Revision: http://reviews.llvm.org/D16218

llvm-svn: 259051