litsupport: Introduce a PREPARE: part for test files.
PREPARE: allows to specify steps that should happen before the benchmark
runs. While you should generally try to keep preparation steps to the
build phase of the benchmark, it is sometimes necessary to remove the
previous results of a benchmark before it can be run again.

We need a PREPARE: step so we do not measure the runtime or attempt to
merge profile data for these steps.

llvm-svn: 269246