[test-suite] Update litsupport/module/microbenchmark.py to report individual timing results from 1 test.
This updates the litsupport microbenchmark plugin module to make further use of
the benchmark library and the changes to lit in "rL327422: [lit] - Allow 1 test
to report multiple micro-test results to provide support…".

I have also changed the current microbenchmarks that use this plugin by
removing the individual tests from the XRay microbenchmarks as they will be
collected individually from the module in 1 run of the executable.

Reviewers: rengolin

Differential Revision: https://reviews.llvm.org/D43316

llvm-svn: 327710