[test-suite] Work around undefined behaviour in lencod
The NUM_BLOCK_TYPES macro is defined to 10, but the arrays in
ctx_tables.h only have 8 elements in the relevant dimension, so this was
causing out of bounds array accesses, and segfaults if we were
(un)lucky.

This is almost exactly the same change as was made to ldecod in D74770.

Differential revision: https://reviews.llvm.org/D77051