[test-suite][AIX] Fix C++ benchmark compatibility issues on AIX
Summary:
This reimplements what has been reverted in 630313d.

The old implementation applied the X/OPEN compatibility flag to the entire miniFE benchmark. This caused a build failure with `mytimer.cpp` due to its usages of `struct timezone` which is not available on AIX when the said compatibility flag is set. This new fix now also defines "UseTimes" which enables another path in the application to avoid the usages of `timezone`.

Reviewers: hubert.reinterpretcast, daltenty, ShuhongL

Reviewed By: hubert.reinterpretcast

Subscribers: mgorny

Differential Revision: https://reviews.llvm.org/D83342