Sampler Lit Tests (#152)
Between the coordinate_normalization_mode , addressing_mode and filtering_mode there are 16 valid combinations that a sampler should support for RGBA images. Here we are adding each combination as its own test.
Some of the devices do not support particular combinations. For example, the CPU device is not correctly supporting linear interpolation at this time. In those cases, I have simply skipped the //RUN: directive for the not-yet-supporting device and left a comment in its place.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>