[test-suite][SPEC2000] Use -fcommon instead of -zmultidef.
This reverts D95276 and D95506 (pass -zmultidefs to the linker) and instead pass -fcommon to the compiler. This is the same solution as already applied to non-SPEC benchmarks in D75557, and caused by a change of the defaults in gcc and clang.

(D75557 adds -fcommon to CPPFLAGS which I think is incorrect because it is not a preprocessor option)

Reviewed By: MaskRay, naromero77

Differential Revision: https://reviews.llvm.org/D98260