Updating the `benchmark` dependency of Microbenchmarks
Updating the dependency to enable leveraging perf counter support.

Renamed the directory to be version-independent. This would allow us to
make updates when needed, without worrying about the relation to the
directory name version. One benefit is that we can avoid cherry-picking
needed changes.

Also added a script, update_benchmark.sh, that automates this - one
needs to just run the script and then submit a patch.

Differential Revision: https://reviews.llvm.org/D101475