[test-suite] Update SPEC2017 CPU cam4 floating point tests to use specpp instead.
Rename the netcdf OBJECT library to cam4_netcdf for consistency with other SPEC tests.

Allow possibility that validator for RATE and SPEED tests might be compiled differently.

Use specpp to process the Fortran90 source files. Split-up compile definitions for specpp processed files and regular files.

Rate test uses an additional compiler definition that needs to be passed to the specpp, but this requires some extra logic to handle properly.

Reviewed By: Meinersbur

Differential Revision: https://reviews.llvm.org/D101846