[benchmark] fix build target timing issues for benchmark
The update to benchmark (D101475) brought in a few new test targets
that need to be built after dependencies such as timeit.

We don't actually need those test targets be built for our purposes, so
we just exclude benchmark tests. This should simplify future rebases of
benchmark, too - tests are likely to evolve, the library target itself,
less so.