[SYCL] Adjust multiple tests to queue::wait() changes (#338)
Adjust tests to expect a call to piQueueFinish rather than piEventsWait
as part of queue::wait(), as well as release of events for queue USM
operations or non-host commands with no implicit dependencies right after
their submission.

Additionally, blocked commands are no longer always waited for before enqueueing
of their users. This introduces more cases where host tasks are used to "glue"
events from different contexts, which makes it so that the order of some calls in the
enqueue-arg-order-image test may vary from run to run.