[test-suite] Only add SOLLVE V&V tests expected to work with Clang.
Some of the test-suite's tests fail with top-of-tree Clang/libomp. Features might not have matured or not supported at all yet (e.g. #pragma omp loop), but it is also possible that tests are not correct according to the specification.

Regardless the reason, this patch isolates the tests that are currently working helping ensuring that those tests are not regressing. All other tests are skipped. The upstream SOLLVE V&V test may add additional tests in the future and Clang add support for existing tests, in which case the list needs to be updated.

Reviewed By: josemonsalve2

Differential Revision: https://reviews.llvm.org/D110047