[test-suite] Fix signed integer overflow error for pc1cod.c
The patch fixes a test fail when compiling the benchmark using GCC
9.3.0. The benchmark was failing because of a signed integer overflow
runtime error in the pc1code.c. This caused the file to have undefined
behaviour and the output generated didn't pass the verification test.

The patch induces the modification of the data type of the variables
causing the signed integer overflow to prevent the error.

Patch by Shrill Shrestha<shrill.shrestha@gmail.com> and Ahana Datta<ahana.datta.ravenclaw@gmail.com>

Differential Revision: https://reviews.llvm.org/D113639