[AIX][test-suite] Use malloc instead of memalign on AIX
This patch adds the _AIX macro to indicate that memalign is unavailable
and should not be used. In these instances, malloc is utilized instead.
Moreover, a new reference output is added for Olden/voronoi as printf on
AIX prints infinites and NANs as INF and NaNQ instead of inf and nan.

Differential Revision: https://reviews.llvm.org/D102409