[SYCL] Extend the SpecConstants/vector-convolution-demo.cpp test (#830)
The test is extended with a specialization constant built upon an aligned
structure. The CFE generates a padding represented as '[size x i8] undef' for
such structures and the sycl-post-link tool crashed during property set
generation for the specialization constant. The bug has been fixed in
intel/llvm#5538, this extension is the E2E test for the fix to ensure that
the default values and offsets given to the runtime are correct.

Signed-off-by: Pavel Samolysov <pavel.samolysov@intel.com>