[Fortran] Forward NO_STOP_MESSAGE to FCVS tests
By default, Flang outputs 'Fortran STOP' as part of the STOP statement,
and sometimes also warnings about signaling exceptions, which are
platform-dependent. This output will cause test failures, but it can be
suppressed by setting the environment variable NO_STOP_MESSAGE=1. This
patch makes sure that llvm-lit forwards the NO_STOP_MESSAGE variable to
the tests if it is set in the environment.

Differential Revision: https://reviews.llvm.org/D126360