[Fortran] Ignore whitespace in FCVS test results
Some tests in the FCVS suite use list-directed output, which may
according to the standard introduce arbitrary amounts of whitespace.
This patch adds the "-i" flag to the fpcmp invocation for those tests,
so that it ignores whitespace. The other tests in the suite will
continue to check whitespace strictly.

Differential Revision: https://reviews.llvm.org/D128260