[CMake][SingleSource] Fix compile with profile generation
Currently, when compiling with the CMake flag
TEST_SUITE_PROFILE_GENERATE set to ON, compare-fp-3.c fails to
compile as clang tries to insert instrumentation on the functions
that are supposed to get optimized out. This commit fixes this by
adding references to these functions when profile generation is
turned on so that the test suite builds without linking errors.

Differential Revision: https://reviews.llvm.org/D128225