[llvm-test-suite][Fortran] Temporarily disable some tests
At the moment, the clang-aarch64-full-2stage buildbot is running the
test-suite with `flang-to-external-fc`, relying on `gfortran` for the
code generation. I'd like to switch it to using `flang-new` instead
(see https://reviews.llvm.org/D130628). However, some of the tests
currently fail, so this patch temporarily skips them until they are
fixed.