[SYCL] Re-enable SYCL/Reduction/ tests (#1153)
reduction_nd_lambda.cpp was supposed to be re-enabled together with
https://github.com/intel/llvm-test-suite/pull/1134/.

For reduction_big_data.cpp the logs with failure expired, not sure how
to deal with it. I did very brief local run of it which passed. As such,
re-enable and deal with it when/if it fails to gather some input data.