[Fortran/UnitTests] add 10 type finalization unit tests
This suite of tests was created  originally by  Wileam Phan, Damian Rouson,
and Brad Richardson as part of the
[[ https://github.com/sourceryinstitute/smart-pointers | Smart-Pointers ]]
library's test suite.
All compilers, except for NAG, did not initially have a working/correct
implmentation of finalization.

The original adaptation for inclusion in the llvm-test-suite can be found here:

* https://github.com/BerkeleyLab/llvm-test-suite/tree/damians-fortran-type-finalization
  SHA: `0268bcf0048e67cd1280f9ef65aebd2aa402130b`
* https://github.com/BerkeleyLab/llvm-test-suite/tree/berkely-lab-damian-v0.1
  SHA: `0268bcf0048e67cd1280f9ef65aebd2aa402130b`

The test suite was then adapted to be made appropriate for inclusion
in a compiler test suite by Izaak Beekman.

A summary of the tests can be found in the README.md file added in
the `Fortran/UnitTests` subdirectory.

Co-Authored-by: Damian Rouson <rouson@lbl.gov>