[SYCL] tests to validate sub-devices have parent (or not) (#1346)
These are the tests for https://github.com/intel/llvm/pull/7167

All devices available when using ONEAPI_DEVICE_SELECTOR are root devices, even those which are gotten via the sub-device selection choices ( e.g. ONEAPI_DEVICE_SELECTOR=level_zero:*.* ). In this PR we are ensuring that all devices gotten with that environment variable are root devices that have no parent. And, conversely, that when not using it, any sub-devices do report the correct parent.