[SYCL] Fix result check in joint matrix tests (#1432)
Some tests in SYCL/Matrix verify their results but only use the results
to print "passed" or "failed". A subset of these tests have checks for
"passed" in the output, but does not use FileCheck to govern the check.
This commit changes these tests to return a non-zero value in the main
function if the result verification failed. This allows the test system
to correctly identify tests that fail this verification.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>