[SPEC2017] Run image validator with TEST_SUITE_RUN_UNDER
Image validators in SPEC CPU 2017 are currently built for the target,
not for the host.  When running a test under target emulation, not being
able to run the corresponding image validator will cause the test to
fail.  This patch enables running image validators using the same target
emulation as passed in TEST_SUITE_RUN_UNDER.

Differential Revision: https://reviews.llvm.org/D139660