[SingleSource/Vectorizer] Add unit tests for conditional scalar assignment pattern.
Dedicated unit tests for loops which contain a conditional assignment inside
the loop body.

An earlier version of this patch was posted as https://reviews.llvm.org/D144909.
Some updates have been made to cover more test cases.

I tested this patch with the CSA patch applied (I will add a link once the CSA
patch is posted) using the following options:
```
-O3 -mllvm -enable-csa-vectorization -mcpu=sifive-x280
```
I also ran it with `-mllvm -force-tail-folding-style=data-with-evl` appended
to the previous option set as well as `-mllvm -force-tail-folding-style=none`.

I also verified that vector code was generated for functions that we are
currently able to vectorize using objdump. This patch contains some examples
that we are not able to vectorize today but should be able to in the future.