[NFC][Py Reformat] Reformat python files in llvm-test-suite
This is an ongoing series of commits that are reformatting our Python
code.

Reformatting is done with `black`.

If you end up having problems merging this commit because you have made
changes to a python file, the best way to handle that is to run git
checkout --ours <yourfile> and then reformat it with black.

If you run into any problems, post to discourse about it and we will try
to help.

RFC Thread below:

https://discourse.llvm.org/t/rfc-document-and-standardize-python-code-style

Specific to this change:
* Does not change third-party code in `MicroBenchmarks/libs/benchmark`.

Fix code to make it parseable by python3/black in the first place:
* Fixed space/tab confusion in `CollectDebugInfoUsingLLDB.py`.
* Change to python3 `print` syntax: `ABI-Testsuite/sample.py`,
  `ABI-Testsuite/test/lit.site.cfg`, `tools/get-report-time`

Differential Revision: https://reviews.llvm.org/D151914