[Fortran/gfortran] Enable "compile" tests in gfortran torture tests.
The tests are compiled along with the rest of the tests at build-time. Some of
the "compile" tests are expected to raise an error. To work around this, at
build-time, the compiler is invoked through a wrapper script which saves any
diagnostics emitted by the compiler and always returns success. If the
compilation succeeded, any object files produced are deleted immediately.

At test-time, the diagnostics saved at build-time are compared against the expected
output.

Differential Revision: https://reviews.llvm.org/D149116