[Fortran/gfortran] add mechanism to track support for alternate flags
The idea is to add feature gated denylist files to be appended to the
base denylists. The additional features are specified with cmake
-DTEST_SUITE_FORTRAN_FEATURES="FEATURE1 FEATURE2". The must be
DisabledFilesFEATURE.cmake files present.

One alternative would be to make including the
DisabledFilesFEATURE.cmake files optional. I decided not to do this
because on my system cmake takes a few seconds longer to configure when
the files to be included are optionally absent. Furthermore, it is
good to get some kind of error if the feature name is misspelt.

As an example, this patch includes denylists for HLFIR. I run with
-fstack-arrays -O3 -mcpu-native -flto -flang-experimental-hlfir, but so
far as I know the additional failures are all due to HLFIR. I would
expect the same denylists to work for just -flang-experimental-hlfir.

I expect there are many tests which are disabled by default which will
now work with HLFIR, but I haven't yet had chance to determine which
these are.

Differential Revision: https://reviews.llvm.org/D156130