[Fortran][gfortran] Disable is_contiguous_1 that is nonconformant
is_contiguous_1.f90 explicitly says that it is testing for an
IS_CONTIGUOUS result with zero sized array that goes against
what is required in the standard. Zero and one sized array are
contiguous, even when obtained with triplet strides that are not
ones.

Flang behavior will be standard conformant and align with ifort
and nagfor after https://github.com/llvm/llvm-project/pull/68869.

So this test needs to be disable.
Move is_contiguous_3.f90 under the same comment since it is failing
for the same valid reason (except IS_CONTIGUOUS was folded at compile
time and therefore already correct with flang).