[MicroBenchmark,LoopInterleaving] Re-land - Check performance impact of Loop Interleaving Count with varying loop iterations (#56)
This microbenchmark attempts to find the impact of loop interleaving count on loops with low trip count for different types of loops, such as loops with or without reduction inside it, loops with or without vectorization inside it.
Note: A subset of test cases run on default. ALL_LOOP_IC_TESTS needs to be set for the whole set of tests.