[llvm-test-suite/utils] Add a simple not tool to the test suite
Add a tool similar to LLVM's not utility which inverts the return code of a
command. The major difference is that the utility is not dependent on LLVM's
libraries and performs less error checking than LLVM's not. It supports the
 --crash option when the command is expected to crash.

The tests in Fortran/gfortran have been updated to use this.