enable execute_command_line tests (#82)
build and pass all three test on Linux.

This is because `execute_command_line` calls `std::system`, and on Windows that is cmd. cmd doesn't have `ls`, the Windows equivalent command is `dir`, and for `timeout` for `sleep`. It will compile and run correct based on the standard, however, the printout from the execution won't be the same on Linux.
For example: https://github.com/llvm/llvm-test-suite/blob/394a62ec2d23a26600062df47a7614201aea43c7/Fortran/gfortran/regression/execute_command_line_1.f90#L10
This will compile but will cause:
```
fatal Fortran runtime error(.......regression\execute_command_line_1.f90:31): Invalid command quit with exit status code: 1
```
as this is expected based on the standard, if cmdstat is assigned with non-zero value but the `CMDSTAT` variable is not present, error termination is initiated.
https://github.com/llvm/llvm-project/blob/c9a6e993f7b349405b6c8f9244cd9cf0f56a6a81/flang/docs/Intrinsics.md?plain=1#L905