[AArch64][FMV] Add a non-comprehensive test for ACLE FMV (#87)
This covers the ones we currently have enabled via apple.inc in compiler-rt, as well as a couple that are not. Each test uses Function Multi Versioning to check for the presence or absence of each given feature, and subsequently exercises that feature if it is presumed to exist. For features that are not thought to exist by the FMV infra, it forks and attempts to try the feature anyway, expecting the resulting thread to die, thus affirming the inverse.