[Fortran] Enable tests that failed when using libpgmath on AArch64 Linux (#93)
Now that we're not using libpgmath, these tests pass on AArch64.

They may work on Windows now but I don't have a build setup to check yet,
and none of the builders run the test suite yet. If/when I can confirm,
I'll come back and remove the Windows check too.