[Fortran/gfortran] Switch to using the static test configuration files (#97)
Switches the existing build system to use the static test configuration
files that were added to the repository in previous commits. The previous
method is jettisoned in its entirety. Several tests have been disabled in the
process. Some of these are because of limitations in the build system which
will be fixed in the future. Others ought to have failed earlier but did not
because they were not being built with the correct compiler flags.