Partially undo gcc-c-torture refactoring
GCC has the notion of "default test behaviour", which is stored in the
`dg-do-what-default` global variable. It specified what should happen
when there is no `dg-do` comment found in the test suite.

Our current behaviour is that if there's no `dg-do`, then we just
compile the test. But for the `gcc.c-torture/execute` directory, the
default behaviour is to run the test, because the `execute.exp` script
manaully sets it so:

```
set saved-dg-do-what-default ${dg-do-what-default}
set dg-do-what-default "run"
gcc-dg-runtest [lsort [glob -nocomplain $srcdir/$subdir/*.\[cS\]]] "" "-w"
set dg-do-what-default ${saved-dg-do-what-default}
```

We should probably model this behaviour in our CMake scripts, but for
now undo the refactoring in the execute/CMakeLists.txt script, and for
tests in that particular directory we always try to run the file.