[Fortran/gfortran] Disable test that requires -std=f2008
Since https://github.com/llvm/llvm-project/pull/87501 flang reports
the correct error code which causes this test to fail loudly.

It always was failing because the test does not work with -std=f2018
which is flang's default (and only) option.

I've confirmed that with gfortran-13 the test passes with -std=f2008
and fails (without output matching flang's) with -std=f2018.