[flang] Disable three error tests (#122)
Some interoperability checks are being relaxed in an upcoming
change to the compiler, and an error has been downgraded to a
portability warning.  This will cause these three tests to
"fail" due to lack of a fatal error, so I'm disabling them
now to prevent build bot failures.

If I knew how to run these tests with "-pedantic -Werror",
I would do that instead.