[AArch64] add a target guard to FMV test
The FMV test uses FMV feature names (e.g. flagm2) which are not
available to -march and therefore arguably should not be available to
the target attribute. This change adds a way to specify the string for
the target attribute independently from the FMV feature.

This is necessary for 2cf1439 (#94279) which had to be reverted because
it caused this test to fail.