[sollve_vv] Autodetected native GPU (#129)
Using the CMake option `-DTEST_SUITE_SYSTEM_GPU=native` or (`-DTEST_SUITE_SYSTEM_GPU=auto`), automatically pre-selects the tests that are expected to work based on the system's GPU.

Also fixes a lot of other issues:
 * `SYSTEM_GPU` violates naming conventions
 * `SYSTEM_GPU` is not a cached variable, i.e. after re-running CMake (e.g. after changing a `CMakeLists.txt`), its value is forgotten
 * Remove `intel` and `regression` options. There are no lists for these
 * Remove possibility to specify multiple pre-select lists. The union of those would run, causing them to fail if run on the other vendor's GPU.
 * Remove duplicate `test_metadirective_arch_is_nvidia.c` in preselect lists
 * Update documentation
 * Add messages whenever SOLLVE V&V is skipped
 * Do not dump test directories into the builddir root
 * Replace `TEST_SUITE_FORCE_ALL` with `-DTEST_SUITE_SYSTEM_GPU=all`
 * Remove uses of `remove_definitions`. Just don't add those flags if you don't want them.

The variable `SYSTEM_GPU` is also used in External/HeCBench and External/smoke. These are based on sollve_vv's previous `CMakeLists.txt` and are not affected.