[HIP] Add a correctness test for the memmove intrinsic
So far, the lowering of device-side memmove intrinsics in HIP (and the AMDGPU
backend) is only tested with syntactic regression tests. This patch adds
functional correctness tests for the device-side memmove intrinsic with and
without overlapping source and destination ranges. By testing various
statically known or unknown move lengths, the different lowering mechanisms for
memmove in the AMDGPU backend are covered.