Fix out-of-tree execution of lit
The lit substitution "%b" is used to find previously compiled helpers
such as fpcmp-target; they're more likely to be in the test source
directory (where the CMake artifacts are) than in the test exec
directory (where lit is run). Note that the two are equal in the default
site configuration, so this change matters only in environments with a
custom site configuration file.