Fix spurious errors in halide tests
We occasionally see spurious llvm-test-suite failures with the
following error message:

    CMake Error: failed to create symbolic link '/tmp/tmp.x1PxF3OhCA/Bitcode/Benchmarks/Halide/local_laplacian/../images/rgb.bytes': File exists

The reason for this is that two halide tests try to copy a test
input to the same location, and create_symlink can fail if these
calls race.

Fix this by putting the inputs for the tests into different
directories, by not making ".." part of the result directory.