[FMV][AArch64] Fix UPASS when "SSBS not fully self-synchronizing".
When running try_ssbs2() on hardware which is affected by the "SSBS
not fully self-synchronizing" errata, the linux kernel mutes the
detection of ssbs2 via hardware caps. As a result the default version
ends up running the ssbs2 code which was expected to trap originally.

To work around this UPASS I have created a whitelist of functions
which are expected to return a different status result.

This fixes https://github.com/llvm/llvm-project/issues/109304.