Fix tools tests on Windows
The commit 89fcc5e introduced some tests for the tools in the
tools/ subdirectory that are always run, but these tests were not
working on Windows due to the executables ending in .exe and one of the
tests being written in bash.

This commit fixes those issues by directly using the $<TARGET_FILE>
generator expressions to get the executable file names, and rewriting
the bash test in python.