[Fortran/gfortran] Initial support to override DejaGNU annotations
The DejaGNU annotations in the gfortran test suite sometimes need to be
overridden, for instance in cases where the one of the compilers generates a
compile error for some non-standard extension while the other does not. This
PR adds support to override the some annotations, such as forcing a test to
pass with flang where it would fail in gfortran.

For now, only some of the behavior can be overridden, but this could be
extended in the future. The documentation was updated to reflect this (some
cleanup of the documentation was also carried out). Similarly, the static
test configuration generation script was updated and cleaned up.

An override file is also provided for two tests that were disabled in
27a6a3a5704 since they are expected to pass in flang. The static test
configuration was also updated to reflect this change.