Add ffmpeg as potential external project
To add the external source, run
"git clone -b release/7.1 https://git.ffmpeg.org/ffmpeg.git"
to clone the ffmpeg sources in the
llvm-test-suite/test-suite-externals directory, or configure with
TEST_SUITE_FFMPEG_ROOT pointing at a directory containing this source.

(As this points to a branch, not a tag, it may be required to check
out the known matching commit 10aaf84f855dbcedb8ee2e3fce307e9b98320946.)

This builds two targets; the ffmpeg command line executable (which
isn't executed as a test) and the ffmpeg checkasm test executable
(which runs compiler generated functions and compares them with
handwritten assembly versions of them). The checkasm execuable
can also be run manually to microbenchmark functions, e.g.
"External/ffmpeg/ffmpeg_checkasm --test=vp9dsp --bench=vp9_avg_8tap_smooth* 0".

It is not very meaningful to benchmark the execution of the whole
checkasm executable, as it runs a different numbers of functions
depending on the number of SIMD extensions available on the target
CPU.

(Benchmarking on aarch64 on Linux currently requires direct access to
the pmccntr_el0 register. To currently use a different
timer register, edit ffmpeg/libavutil/aarch64/timer.h and change it
to use cntvct_el0 instead of pmccntr_el0 on Linux too.)

This uses a static configuration of the project (when building the
upstream project with their own build system, there's a number of
build options that can be configured). Assembly is hooked up and
enabled on i386, x86_64, arm and aarch64.

For architectures other than those, the checkasm test won't have
any reference for detecting e.g. miscompilations of functions,
but building it can still be meaningful (for testing compilation,
or benchmarking the execution of the C version of functions).

Instead of having the CMake configuration check for a large number
of things and generate a config.h file at configure time, things are
either hardcoded, or set up with preprocessor conditions within
the hand-tweaked config.h file.