[Fortran/gfortran] Disable tests on some platforms, but enable them elsewhere
Some tests fail only on certain platforms, but pass on others. Previously, these
were disabled everywhere, but now they can be disabled only on the platforms
on which they fail and enabled elsewhere. In addition, some tests that were failing
only on AArch64 now pass, so they are enabled everywhere.

The static test configuration has been updated to reflect these changes.

Some bugs were exposed in the update script during this process. These have
been fixed. Primarily, * is now allowed as a wildcard in the override.yaml file.
Earlier, this would not work correctly since * is not a valid matcher in CMake
regexes. This is now correctly translated to the equivalent .+ in CMake.