[Fortran] disable teams-3.f90 (#211)
This test is expected to fail but now passes. It previously failed
because it triggered a not-yet-implemented error, not because it failed
for the reasons intended by the test.

The test checks that some openmp library calls are not allowed inside of
omp teams. These semantic checks are not currently implemented in flang.