[Fortran] Add floating point to integer conversion test (#220)
Semantics for floating point conversions to integer types was changed in flang
to use the saturated floating point intrinsics to more closely match the
standard semantics in llvm-project#130686 .
This tests conversions from tiny, huge, nan, and inf to various integer types.
Tests for 128b types are enabled if the fortran compiler supports it.