Remove pre-compiled binaries
For some reason, we have three pre-compiled binaries checked in for
the gs, agrep and football benchmarks. These are not actually used
(we'll build our own, as usual). Presumably they were committed by
accident.

We also ship one more pre-compiled makeproto binary for the p2c
benchmark. This is not necessary either, as we already ship the
`.proto` files it generates.