[Fortran] disable pointer_check_11.f90 which exhibits UB (#232)
This test calls an empty subroutine with a null pointer as an argument.
In gfortran it is hoped that a flag will detect this and produce an
error. Flang does not perform any such analysis, and so on some machines
when optimization is enabled, this test never terminates. I don't think
this is a flang bug as such because the input code is doing undefined
behaviour.