[Fortran/gfortran] Sync gfortran tests with upstream (#241)
These are synced with c182f4d14d65b3e012.

Any new tests that fail have been added to the appropriate deny lists. The
reasons for these failures have not been thoroughly investigated though, so
some recategorization may be necessary.

Some bugs in the build system for the tests was uncovered in the process and
fixed.