[Fortran/gfortran] Skip some directories when generating test configurations (#246)
Some directories in the gfortran test suite only contain "dependent" sources.
These are sources used in multi-file tests where only the "main" source file
contains DejaGNU directives that are used to determine the behavior of the
test. A list of such "dependents-only" directories have been added to the
update-test-config script so a test configuration is not generated for them.