[Fortran/gfortran] Disable failing test
This test should fail to compile with an "invalid character in name" error - or
the equivalent in flang. However, it used to fail because the LINEAR clause was
not implemented in the OpenMP DO construct. The clause is now support. However,
flang does not issue the expected compile-time error which causes the test to
fail.